Restruct. config related things
